Refactor LazyGetter
